+++ gcms2isocor
@@ -1,2 +1,2 @@
 #!/bin/bash
-java -jar target/scala-2.13/P2M2WorkflowsTools-assembly-0.1.jar $*
+java -cp target/scala-2.13/P2M2WorkflowsTools-assembly-0.1.4.jar fr.inrae.metabolomics.p2m2.GCMS2IsocorCommand $*
