+++ openLabCds2Csv
@@ -1,0 +1,2 @@
+#!/bin/bash
+java -cp target/scala-2.13/P2M2WorkflowsTools-assembly-0.1.4.jar fr.inrae.metabolomics.p2m2.OpenLabCDS2CsvCommand $*
