+++ src/test/resources/OpenLabCDS/Results.csv
@@ -1,0 +1,4 @@
+;Glyoxylate;Glycolate;Succinate;Glycerate
+EX1;10.97505;19.37573;22.58038;27.60959;
+EX2;11.20124;33.88725;58.95049;34.04986;
+EX3;13.78376;44.91168;110.09505;158.98767;
